perf: isolate timer state into TipRotator to prevent parent re-renders
